move GuessBaseImageIndex to `pkg/fanal/image.go`
