+++ pkg/fanal/image/image.go
@@ -2,6 +2,7 @@ package image
 
 import (
 	"context"
+	"strings"
 
 	"github.com/google/go-containerregistry/pkg/name"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
@@ -127,4 +128,32 @@ func LayerIDs(img v1.Image) ([]string, error) {
 		layerIDs = append(layerIDs, d.String())
 	}
 	return layerIDs, nil
+}
+
+func GuessBaseImageIndex(histories []v1.History) int {
+	baseImageIndex := -1
+	var foundNonEmpty bool
+	for i := len(histories) - 1; i >= 0; i-- {
+		h := histories[i]
+
+		// Skip the last CMD, ENTRYPOINT, etc.
+		if !foundNonEmpty {
+			if h.EmptyLayer {
+				continue
+			}
+			foundNonEmpty = true
+		}
+
+		if !h.EmptyLayer {
+			continue
+		}
+
+		// Detect CMD instruction in base image
+		if strings.HasPrefix(h.CreatedBy, "/bin/sh -c #(nop)  CMD") ||
+			strings.HasPrefix(h.CreatedBy, "CMD") { // BuildKit
+			baseImageIndex = i
+			break
+		}
+	}
+	return baseImageIndex
 }
